Implemented DM misregistration. Apply registratrion to both DM and WFS to model unknown pupil misregistration. No need to apply pupil misregistration to science.

+++ lib/misc.h
@@ -19,7 +19,7 @@
 #ifndef AOS_LIB_MISC_H
 #define AOS_LIB_MISC_H
 #include "common.h"
-long factorial(long n);
+double factorial(long n);
 char *mybasename(const char *fn);
 int check_suffix(const char *fn, const char *suffix);
 void copyfile(const char *dest, const char *src);
